Move derive tests out of rex-engine and into rex
In commit 38a6fd6 I added some tests for ADTs that depend on the
derive(Rex) proc macro, but I put them in rex-engine which meant that
rex-engine had to depend on rex-proc-macro. I only realised after that
there was already a place for tests that use derive in the rex crate
itself. This commit corrects this.

Since these tests rely on parse_infer_and_eval() which was previously
only used by the other eval tests, this commit now moves that function
to a separate file, rex-engine/src/util.rs, not gated by cfg(test), and
exposes the function publicly so it can be used in the derive tests in
the rex crate.

These changes are a necessary pre-requisite for the next commit in which
the derive(Rex) proc macro is modified to reference the symbols exposed
from the rex crate, rather than rex_type_system, which is necessary for
it to be made available by tengu, which only depends on the main rex
crate.