Builder: Add register_fn_raw()
This is a convenience mechanism for registering functions that need to
deal with Expr types as inputs and outputs, rather than Rust types.

Functions like this are necessary in cases where the number of arguments
and their corresponding Rust types cannot be known in advance, such as
when registering module handlers in tengu.