Avoid duplicate ids in try_enocde()
Each expression is supposed to have a unique id, so that it can be used
as a key to look up its associated type in ExprTypeEnv. If two or more
Exprs have the same id, this lookup could return the incorrect type for
some of them.

The various implementations of Encode were passing the same id down to
child elements (e.g. for tuples), which would cause the above to occur.
Actually try_encode() should not accept an id at all - it should instead
generate its own unique id by calling Id::new() for each Expr that it
creates.