fix: derive(Rex) consistency with register_adt
When encoding structs in derive(Rex), wrap these in Named Exprs in the
same way that we were already doing for enum variants.

This is necessary to make the Encode/Decode implementations generated by
derive(Rex) consistent with the behaviour of the constructor functions
created by register_adt. The latter has no way to distinguish between
structs and enums with a single variant, because of its use of ADT
rather than the Rust parse tree from the syn crate.