feat: serde_json-compatible JSON serialization
Add functions to convert between Exprs and JSON values in a manner that
is compatible with serde_json. This is necessary for exchanging data
between rex and tengu modules.

These functions differ from the Serialize and Deserialize derived traits
on Expr in that the latter includes the id and span, and uses different
ways of representing named data types, optional values, and other kinds
of values. These representations are not understood by the modules.