feat: Extend register_adt with prefix, defaults
Allow a prefix to be supplied for use when forming the names of
constructor functions for the variants of an ADT. This is necessary in
cases where you have a Rust enum with one or mre variant names that
overlaps with a Rust struct or other enum that you have also registered.

Also allow a map of defaults to be supplied, which are fields that will
be filled in by invoking a callback function when an instance of the ADT
is created. Callback functions are used so that they can obtain
information from the context - such as the current date & time when
initialising a "created_at" field.