perf: Avoid mutating types
In the next commit we'll switch to using Arc<Type> everwyhere, which
will prevent us from being able to treat types as mutable. The only
place where this is actually done is in resolve_vars(), which is called
in a single location, register_fn_core(). To prepare for the Arc change,
make resolve_vars() return a new Type instead of modifying the existing
one.