perf: Use Arc<Type> everywhere
There are several places within both the type inference and evaluation
code where we cloned types; this is expensive when there are a lot of
ADTs registered or the are large substitutions.

The most significant performance impact of expensive type cloning can be
seen in unify_one_of() when there are multiple overloaded functions
registered with a given name, such as is the case when multiple ADTs are
registered with the same field names. The num_fields benchmark measures
this.

By using reference counting instead of cloning, we drastically reduce
the time spent in unify_one_of(); in the case of the num_fields
benchmark, the speed improves by a factor of 4. Reference counting
likely helps in other places, though these have not been measured.

Before this change:

$ cargo run -- benchmark --name num_fields
compile_time with 100 fields, 150 entries = 12903 ms

After this change:

$ cargo run -- benchmark --name num_fields
compile_time with 100 fields, 150 entries = 3196 ms