perf: Minimise no of constraint solving iterations
The constraint checker needs to loop multiple times, because it is often
the case that changes to the substitution that happen as a result of
unifying later constraints in the list make it possible for changes from
those earlier in the list to become possible to solve.

Previously, we used a hard-coded count of 100 iterations. This commit
changes that so it keeps track of whether the substitution has been
modified in a given iteration, and stops once no further changes have
been made.