fix: Don't treat int enums specially
In the derive(Rex) macro, don't make any distinction between enums with
int values and those without. The reason we were doing this was due to
me incorrectly assuming that the corresponding derive(TypeInfo) macro in
ouroboros treated such enums as Uint, when it turns out this isn't the
case.