chore: Avoid separate Subst type for change checks
When performing unification, we need to keep track of whether or not a
change has been made to the substitution to decide whether we need to do
another iteration on the list of constraints. Rather than having a
separate type for the substitution which includes this field, which is
only needed for the unification process, pass the field as a parameter
to the functions that set it. This avoids exposing this internal detail
of the algorithm to code that doesn't need to know abouot it.

One caveat to keep in mind is that if making changes to the algorithm in
the future, it's critical to ensure the flag gets sert wherever the
substitution is modified, now that it is tracked as a separate variable.
For example this needs to be done in unify_one_of().