chore: Include ToType/Encode/Decode in JSON tests
Start combining the various tests for the derive(Rex) macro and JSON
conversing. These tests have a lot of overlap, and to make sure each
kind of operation is implemented consistently, it seems best to test
them all in one place for a given type defintion. The operations we seek
to test are:

- ToType::to_type()
- Encode::try_encode()
- Decode::try_decode()
- expr_to_json()
- json_to_expr()