fix: JSON conversion for empty structs
Continue combining the derive/encoding/json tests; in this case this
surfaced a bug where empty structs (those with no fields) were not being
encoded and converted to JSON in a manner consistent with serde.
Although this case was tested in derive.rs, this missed the JSON
conversion logic which is why this bug was missed previously.