feat: Type::transform() and Type::for_each()
The transform() method traverses a tree of types and calls the supplied
function for each node; if the function returns Some(t), then the type
is replaced with t in the result. If the function returns None,
traversal continues.

This function is used for cases where you need to replace certain
subtypes, e.g. replacing Object<T> with VirtualObject in tengu.

The for_each() method is a convenient way of visiting all nodes in a
type tree and can be used for things like identifying the set of all
ADTs referenced within a type.