feat: Prefix enum constructors with ADT name
Avoid variant name conflicts between different ADTs by prefixing the
constructors with the ADT name itself, e.g. Color::Red.

This is only done for ADTs that correspond to enums. ADTs that
correspond to structs (those with a single variant whose name is the
same as the ADT itself) do not have the prefix added.