chore: Use sync RwLock for type environment
In all cases where we access the type environment (which is shared
between different async tasks), we only do so to get or insert an entry,
both of which are trivial operations. We don't need to hold the lock
across await points, so there is no advantage to using tokio's version
over the Rust stdlib implementation.

See https://docs.rs/tokio/latest/tokio/sync/struct.Mutex.html under the
section "Which kind of mutex should you use?" (this also applies to
RwLock).