chore: Helpers for accessing type environment
Add the get_type() and set_type() methods to Context, to simplify
relevant parts of the evaluation code and centralise all access to the
type environment in one place.

The get_type() method applies the current substitution to the type, as
was being one already in all the places it is used.

These methods hide the fact that they are accessing shared state, but
still do so safely, as they are doing so using RwLock.