fix: Avoid panic on expr type lookup
The fact that there are Exprs for which we do not have the type recorded
is a bug, and is not something we should ever encounter when the
evaluator is functioning correctly. Until then, return an error rather
than panicking, so that when running inside tengu it doesn not kill the
engine.