fix: Disallow overloads w/different param counts
When registering a function (whether its an accessor or a normal
function), report an error if there is already a function in the ftable
with the same name but a different number of parameters. All overloaded
functions have to take the same number of parameters due to the fact we
allow for currying.

While this check is useful in itself, it's primarily intended for an
upcoming change in which we determine the number of arguments required
for a function by looking up its name. For this reason, the function
table contains Entry structs, each of which has a num_params field.