fix: Implement dynamic dispatch w/exprs, not types
This commit fixes two bugs relating to dynamic dispatch in the apply
function. Until now, dynamic dispatch has operated by keeping track of
the type of every expression, and when calling a function, deciding
which implementation to invoke based on the type of its arguments. This
required careful maintenance of the type environment at runtime to
ensure that the correct types were recorded for every expression, and
there were cases where this was being done incorrectly.

While these could have been fixed, I relised there's a much simpler
solution: if we consider what information we need to choose which
implementation of a function to call at runtime, then for all the cases
we are about, we can actually do this based purely on the argument Exprs
themselves. That is, we don't need to track the types at all. Overloaded
functions that accept primitive types (arithmetic operators, ==, etc.)
are trivial since Expr enum has separate variants for Uint, Int, Float,
etc. Overloaded functions for more complex types can be distinguished by
the values of their expressions - e.g. Option and Result can be
distinsuished by looking at the variant name in a Named expression
("Some" or "None" for Option, "Ok" or "Err" for Result).

The only case of relevace to our use cases is ADT accessors - these are
of the form A -> B where A is the type of the ADT, and B is the type of
a particular field. For example, the "account_id" accessor function has
the type Protein -> String. The key thing however is that all accessors
for a given key have the same implementation - they simply look up an
entry in a dictionary - so they can be reused. So for all accessors of a
given property name, we only need a single implementation - though for
type checking purposes we still act as if each instance of an accessor
has a separate type.

Concretely, here's what has changed:

The type environment in Context is gone; we don't keep track of types at
runtime. Instead, when we need to decide which implementation of a
function to call, we make that choice solely on the value of the
expressions supplied as arguments ot the function. This is done using
the maybe_accepts_args() function of the Dispatch trait, which is
implemented by both Type and Accessor. For "normal" functions, the
dispatcher is the type of that function; for accors, its simply an empty
struct that returns true for any named expression.

Variable lookup has been simplified; eval_let() evluates definitions
strictly instead of deferring them, since it can do so without needing
to know the type in advance, and eval_var() simply returns the Expr
associated with the given name, or a Curry referencing the appropriate
built-in function in the case where there is no let- or lambda-defined
variable with that name.

Curry expressions are used to represent all references to functions;
eval_app() only needs to handle Lam and Curry Exprs, not Var, since the
latter is already resolved to one of the former by eval_var().

The code in eval_app() that was updating the type environment for lambda
expressions has been removed, since it is no longer needed.

The one casualty of these changes is that we no longer support
polymorphic functions of zero arguments, since the dynamic dispatch
mechanism now relies only on the arguments, not the expected result
type. The "zero" function and related tests have thus been removed. If a
case comes up where we need something like this we'll have to find
another way to handle it; having said that, the None function (which
returns an Option<T>) still works, by virtue of the fact that any
function requiring an Option<T> to be supplied as its argument will
accept the named Expr "None".