feat: Overloaded polymorphic functions
Make it possible to register multiple built-in functions with the same
name which are generic (that is, they contain type variables). Examples
include map, and_then, or_else, unwrap_or_else, and the elem{N}
functions for accessing tuples. The way to make this work became
apparent while making changes necessary to report location information
for type errors (see next commit).

Reporting location information for type errors requires us to store a
Span in each constraint. However, when adding constraints at function
registration time, the expression tree is not yet available, so we
cannot assign a meaningful span. An initial attempt was made in which
the spans were optional, and while this gave location information for
some errors, there were many instances where the constraint associated
with a type error would be one of those created during function
registration. This lead me to the conclusion that we need to avoid
creating constraints altogether until the expression tree is available.

This commit makes three key changes to the way types are handled at
function registration time:

- Instead of generating OneOf constraints for overloaded functions, we
  instead build a type environment associating the name with a *set* of
  possibilities. For functions that are not overloaded, this is simply a
  set of one element.

- Each overload is represented as a type scheme, rather than a type. The
  difference between these are that type schemes are quantified and must
  be instantiated before use; types are not. This subsumes the
  functionality of Type::ForAll and avoids any ambiguity about when
  instantiation is necessary.

- Unnecessary equality constraints, which were the source of spanless
  type errors in the initial attempt above, are avoided since we
  postpone generation of these until variable lookup time.

When traversing the expression tree in generate_constraints() and
encountering a variable, we look up its name in the type environment to
get the set of possible type schemes it can have. If there is only one,
this means the function is not overloaded, and we simply instantiate it
to get the variable's type. If there are multiple, we instantiate all of
them, and generate a new OneOf constraint with a fresh type variable.
The result from this point forward is equivalent to what was generated
previously, except that *all* constraints now have spans associated with
them by virtue of being created only during traversal of the expression
tree.

A consequence of the fact that we now treat overloaded and polymorphic
functions in exactly the same way at function registration time is that
the problem of overloading the latter simply disappears. The one thing
we do need to check for however, to avoid runtime errors, is to make
sure the types of all functions with a given name are disjoint. This is
handled by the maybe_overlaps_with() method in Type, which operates
analogously to the maybe_compatible() method used at runtime but with
types instead of expressions.