feat: Report location information for type errors
Add a new error type for type checking, which contains the span
corresponding to the expression to which the error relates. These spans
are stored in the constraints generated during initial traversal of the
expression tree.

Most of the work required for these changes was done in the previous
commit; this one just adds the spans and errors.