feat: Report multiple type errors, not just first
If we encounter an error during type checking, record it and keep going,
so that we can catch as many errors as possible. This helps speed up
porting of code by allowing all the problems to be displayed in one go,
instead of fixing one error at a time and re-running the type checker.