feat: Report multiple errors for unification
When performing unification, keep going if we encounter an error, and
build up a list of all errors found during the unification of two types.
Now that we have paths included in the error messages, it's made clear
which error relates to which property.

This is useful when calling an ADT constructor and there are errors with
multiple properties.