feat: Encode/Decode support for serde_json::Value
Some Rust types we need to use with Rex have fields of type
serde_json::Value. Although we cannot meaningfully work with JSON in
typed rex (at least currently), we at least need to not fail when the
derive(Rex) macro is used on these types, and preferably be able to
actuallly round-trip these values when converting from the rust type to
Expr.

This commit implements the Encode, Decode, and ToType traits for
serde_json::Value. It encodes these as named values containing the
string representation of the JSON value.

Since json_to_expr() and expr_to_json() are supposed to exactly match
the behaviour of serde, we also explicitly recognise encoded JSON data
in this expression format in these functions.