feat: Use reference counting for Expr
Cloning Exprs happens a lot and can be very expensive, particularly when
a given expression is used many times such as in a call to map. Like we
did with types a little while back, wrap all Expr values in Arc, so they
can be shared. Having multiple references to a given expression is now
cheap.

The performance improvements observable in this commit mostly relate to
cases where a function containing lots of code is called many times,
such as mapping over a large list. This is because for each call, the
function was previously copied in its entirety; now only a reference is
copied. In the case of benchmark_map_large_function, when run with
list_size = 10000 and function_count = 10, these changes reduce
execution time from 16.5 seconds to 1.3 seconds. The memory savings are
modest (from 18.3 mb to 11.9 mb), since map is (currently) sequential so
only one of the copies of the function applied to map is in memory at a
time.

To *really* make the most of sharing requires one more change - avoiding
copies caused by evals - which will be addressed in the next commit.