feat: Avoid identical copies in eval
When evaluating a list, tuple, or dictionary, it may be that all the
elements have already been evaluated. In this situation, the Arc
references we get back from the nested eval calls will be the same. When
this is true of *all* the values in a collection, we can reuse the
original list, tuple, or dictionary instead of creating a new one with
the same contents.

This is where the Expr reference counting change really shows its
benefits. In the case of benchmark_large_lists, when run with n = 100,
this reduces memory consumption from 516 mb to 21 mb, and execution time
from 24 to 2.3 seconds.