chore: Remove local/global constraint distinction
In the past, we had two types of contraints: global constraints were for
built-in function definitions, while local constraints were those
generated as the result of program analysis.

The former are no longer used, having been subsumed by the changes in
commit 7a274e1, which moved this information into the type environment.

For this reason, we only need to store one list of constraints in the
constraint system.