chore: Support different kinds of types
Change the way we represent parameterized types to match the approach
used by [1], which allows us to use type variables in the position of
type constructors, not just their arguments. This is one of the steps
necessary to support type classes properly.

Types have "kinds", which determine how they may be used. A kind is like
a type of a type. The kind * is a concrete type, such as uint, bool,
[string], or float -> bool, which can be used directly. Types with this
kind are what programmers normally think of when they use the word
"type". Kinds can also represent type constructors, which are the type
equivalent of functions; a type with the kind * -> * accepts one type as
a parameter and returns a concrete type with kind *. This can be
extended to multiple parameters.

For example, List has the kind * -> * (the single parameter is the type
of the items in the list), and Arrow has the kind * -> * -> * (the first
parameter is the type of the function, and the second parameter is the
type of the argument).

Instead of a type of the form T(x1, x2, ...) where T is a Rust enum
variant and x1, x2, ... are types, we now have a "type application" enum
varaint App which can be composed as many types as necessary to make up
a type application, e.g. App(App(T, x1), x2). In the latter case, T
itself is a type, and could be a constant, type variable, or another
application.

The following parametric types supported by Rex now have their
representation changed as follows (with the Arc and Type:: prefixes
elided for brevity):

    Arrow(a, b)   is now   App(App(Arrow, a) b)
    Result(a, b)  is now   App(App(Result, a), b)
    Option(a)     is now   App(Option, a)
    Promise(a)    is now   App(Promise, a)
    List(a)       is now   App(List, a)

Type constants are now members of the TypeCon enum. These include those
listed above like Result and List, as well as atomic types like Uint and
String. Dict and Tuple have been left unchanged as these do not lend
themselves to the parameterization approach discussed above.

The purpose of this change is to allow types to be defined in which
variable type constructors are applied to their arguments, which is
necessary for properly supporting type classes. For example, Haskell has
a typeclass called Functor containing a single method, fmap, defined as
follows:

    fmap :: Functor f => (a -> b) -> f a -> f b

In this type signature, the 'Functor f' part is a predicate that defines
a type variable f, which is required to be a member of the Functor type
class, which requires it to take one parameter (in other words, f has
kind * -> *). The type signature contains two different applications of
f to different parameters, reflecting the behaviour of map. For example,
if f as List, then fmap could be applied to a function of type Int ->
String, yielding a curried function that accepts an argument of type
List<Int> and returns a value of type List<String>.

As part of this change, we have swapped the order of parameters to
Result; we now use Result<E, T> instead of the former Result<T, E>,
where T is the success type and E is the error type. This is so that we
can define map on Result<E, _>, e.g. allowing calls to map from
Result<E, X> to Result<E, Y>. This would not be possible without
changing the parameter order, as it relies on partial application of a
type constructor. It is equivalent to Either in Haskell.

[1] Mark P. Jones. Typing Haskell in Haskell.
https://web.cecs.pdx.edu/~mpj/thih/thih.pdf