chore: Add TypeVar & Kind
More data types from the "Typing Haskell in Haskell" paper.

Record the kind of each type. In the general case, kinds can have a tree
form, where each kind is either * or a -> b, where a and b are kinds.
For our purposes however, we only need to deal with kinds where the
argument is of kind *, so we simply represent the kind as the number of
arguments; 0 means a "concrete" type, which is the default.

TypeVar, TypeCon, and Type all have kind() methods. The latter can
return an error, in the case where an invalid type has been constructed
by applying a type of kind * to an argument, which doens't make sense.
This is indicated by returning either Some for a valid kind or None for
such an error.

TypeVars are used in place of ids. The difference between the two is
that TypeVars have a kind. Attempting to unify two types with different
kinds produces an error.