chore: Separate function wrappers & registration
Until now, all functions have been registered with the builder directly,
by calling a function of the form register_fnX or register_fn_asyncX.
These took a Rust function/closure as an argument, wrapped it in a
FtableFn to handle conversion between Rex expressions and native Rust
types, and associated that wrapper function with the given name in the
builder's function table.

When adding support for type classes, we're going to need additional
ways of registering functions - some will be handled the same way, but
others will be registered on instances of the type class, likely with
extra information (such as the type to which the instance applies). In
the latter case, we'll need separate function(s) for registration.

To avoid code duplication, this commit separates out the logic for
creating the FtableFn wrapper for a given Rust function from the
registration process. This allows us to add additional forms of
registration in the future.