chore: Move several functions into Type
To more closely align our code with that of the "Typing Haskell in
Haskell" paper, move the following functions into Type:

- free_vars
- occurs_check
- apply (formerly apply_subst)

In the paper, these functions are also defined for other data structures
(such as type schemes, as is also the case for us), so having a member
function with a single name makes it easier to match up our code with
the paper. Apart from apply, the actual names we use are different from
the paper, for clarity.