chore: Restrict OneOf constraints to type vars
The previous commit revealed an interesting insight: OneOf constraints
always have a type variable on the left, not an arbitrary type. There
are two cases:

1. When a new OneOf constraint is added, by calling add_one_of(). The
   argument to this always happens to be a type variable.

2. When instantiating a type scheme, and we apply a substitution mapping
   type variables to new type variables to the constraints. This
   preserves the restriction.

This insight is promising because it suggests we may be able to store
OneOf constraints more efficiently, in a map keyed by type variable,
rather than a list. We leave such optimisations to future commits; here
we just change the Constraint enum to reflect the above restriction.