chore: Remove negate overload for uint
For type class support, negate will need to have the type:

    Num a => a -> a

In other words, the result type will need to be the same as the input
type. This works for floats and signed integers, but not for unsigned
integers. Therefore we cannot support negate for unsigned integers while
also supporting type classes.

This restriction also exists in Haskell, for the same reasons.