chore: Monomorphize or_else, unwrap_or_else
These two functions can't be defined in a way that's compatible with
type classes (or at least type classes with a single parameter). Even if
we were to support type classes with multiple parameters, this wouldn't
work for Option since that only has a single parameter.

To avoid this problem, go back to having _result and _option versions of
these functions.

The other generic functions on Result and Option (map, and_then, unwrap)
are fine because they are parameterised on a single parameter, which in
the case of Result is the type of Ok values.