chore: Store errors in constraint system
For consistency with the previous change that moved the substitution
into the constraint system, and to simplify the code a bit, store the
errors alongside the substitution inside the constraint system. The
substitution and errors are closely related, since errors are generated
during both constraint generation and constraint solving.

This doesn't impact performance; it's just for convenience.