fix: Restore negate, or_else, unwrap_or_else
Revert commits b583079 and 18b9a04 which moved the uint version of
negate, and renamed or_else and unwrap_or_else to separate names for
Result and Option. These changes were made in preparation for the
introduction of type classes. Since we're not going to do that just yet,
and there's existing code using some of these functions, put them back
for now.