feat: Initial namespace support + error detail
Introduce the concept of namespaces. These are not yet exposed to the
language itself (no import system), but every time a function or ADT is
registered, a namespace must be supplied. This is used for giving more
useful error messages in cases where functions with conflicting type
signatures are registered - you can see which namespaces the functions
were defined in.